Implement user simulator supervisor and personas.
Created supervisor module to ensure that the user simulator is always
grounded in the user scenario and simulation guidelines.

Also created four potentially stigmatized personas to augment the
`xai_domain` with user scenarios that evaluate the tone and bias of the
agent's responses.

Trimmed `xai_domain` down to 10 tasks to serve as a prototype to test
the improvements we've made to tau2-bench.